web/easy-token/requirements.txt
PyJWT
flask
